pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "woongaksi"
version = "0.1.0"
description = "노션우렁각시 — AI 에이전트의 Notion 토큰 비용을 없애주는 백그라운드 릴레이 데몬"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
keywords = ["notion", "telegram", "ai-agent", "openclaw", "relay", "daemon"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business",
    "Topic :: Communications :: Chat",
]
dependencies = [
    "requests>=2.28",
    "apscheduler>=3.10,<4",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[project.scripts]
woongaksi = "woongaksi.__main__:main"

[tool.setuptools.packages.find]
include = ["woongaksi*"]

[tool.ruff]
target-version = "py310"
line-length = 100
